requirements.txt
google-genai
flask
python-dotenv